Move routing utils to routing folder
